requirements.txt
langchain
langchain-openai
langchain_text_splitters
langchain-community
langchain-chroma
chainlit
python-dotenv
chromadb
torch
transformers
sentence_transformers
unstructured
pypdf
ctransformers
Markdown
tiktoken
openai